CommonBackend is extracted from SQLA and memory backends.
